requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
et-xmlfile==2.0.0
idna==3.11
openpyxl==3.1.5
pytelegrambotapi==4.23.0
python-dotenv==1.0.1
requests==2.32.4
urllib3==2.2.3
